+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Use VecDeque for interval history to avoid O(n) removal on every interval
+
 ## [0.3.0] - 2026-01-31
 
 ### Added